requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-telegram-bot==20.3
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
